Minecraft: update launcher file
Included in the recent beta 1.3 update is the ability to pass to the
java launcher [username, password, server, [port]] to force the launcher
to try to log in right away. We've scripted the launch of the java gui
launcher. This passes the arguments to our script on to the laucher.

Also, exec the whole line -- cleaner.